fix: Fix to insert rows in initiating if not existing (#7)
[Problem]
When request `GET /games/{game_id}`, it initiates Inning and Playing table every time. So they were duplicated every time.

[Solution]
When initiate the tables, it checks whether they are exist or not. And if it does not exist, insert row at the table. (Fixed InningService::initInning and PlayingService::initPlaying. Added InningRepository::findByTeam and PlayingRepository::findByTeam)